продукты

+++ src/ListProducts.java
@@ -1,24 +1,28 @@
 import java.util.*;
 
 public class ListProducts {
-    static Map<Product, Double> products;
+    static Map<Product, Integer> products;
+
     public ListProducts() {
-        products =new HashMap<>();
+        products = new HashMap<>();
     }
-    public void addProduct(Product product) throws ProductException{
-        if (products.containsKey(product)){
-            throw new ProductException("Продукт "+ product.getNameProduct()+ " уже есть.");
+
+    public void addProduct(Product product) throws ProductException {
+        if (products.containsKey(product)) {
+            throw new ProductException("Продукт " + product.getNameProduct() + " уже есть.");
         } else {
             products.put(product, product.getWeightProduct());
         }
     }
-    public void deleteProduct(Product product){
+
+    public void deleteProduct(Product product) {
         products.remove(product);
     }
-    public static Double sumPriceProduct(){
-        double sum =0;
-        for (Map.Entry<Product, Double> entry : products.entrySet()){
-                sum+=entry.getKey().getPriceProduct()*entry.getValue();
+
+    public static Integer sumPriceProduct() {
+        int sum = 0;
+        for (Map.Entry<Product, Integer> entry : products.entrySet()) {
+            sum += entry.getKey().getPriceProduct() * entry.getValue();
         }
         return sum;
     }
@@ -30,15 +34,15 @@ public class ListProducts {
 
     @Override
     public boolean equals(Object obj) {
-        if (this==obj)return true;
-        if (obj==null || getClass() !=obj.getClass()) return false;
-        ListProducts that=(ListProducts) obj;
+        if (this == obj) return true;
+        if (obj == null || getClass() != obj.getClass()) return false;
+        ListProducts that = (ListProducts) obj;
         return Objects.equals(products, that.products);
     }
 
     @Override
     public String toString() {
 
-        return super.toString()+products;
+        return super.toString() + products;
     }
 }
